Replace uses of "PyCon" with "EuroPython" as far as possible.
Change the homepage discover "buttons" to use English text.

Update the footer line to the EuroPython 2015 theme footer.

Fix the name of the background image to read "europython" instead
of "pycon" in the CSS/HTML files.

Add a few warnings for account details which should not be
in a public repo.

Fix website addresses to use "europython.eu" instead of "pycon.it".

Fix newsletter subscription address to point to europython-announce
(this doesn't fully work yet, because Mailman uses a different subscription
interface than Google groups).