Remove rendundant call to publish_page()
As the page is already published, this was causing Django-CMS to fail
with a `NodeAlreadySaved` exception from treebeard. The only thing it
looked to be doing that `cms.api.create_page()` was not was assigning a
user to the page, and this can already be done with `create_page()`
using the `created_by` argument.

Fixex #1349.